revert useless change (GBI_FLOATS is defined)

+++ include/PR/gbi.h
@@ -1183,7 +1183,7 @@ typedef struct {
  * First 8 words are integer portion of the 4x4 matrix
  * Last 8 words are the fraction portion of the 4x4 matrix
  */
-typedef f32	Mtx_t[4][4];
+typedef s32	Mtx_t[4][4];
 
 typedef union {
     Mtx_t		m;
